refactor: move BlogIndex to GenericPagesViews
