+++ src/__tests__/pages/teachers/curriculum/index.test.tsx
@@ -8,7 +8,7 @@ import CurriculumHomePage, {
 import renderWithProviders from "@/__tests__/__helpers__/renderWithProviders";
 import subjectPhaseOptions from "@/browser-lib/fixtures/subjectPhaseOptions";
 import SubjectPhasePicker from "@/components/SubjectPhasePicker/SubjectPhasePicker";
-import { SerializedBlogPostPreview } from "@/components/pages/BlogIndex.page";
+import { SerializedBlogPostPreview } from "@/components/GenericPagesViews/BlogIndex.page";
 
 const render = renderWithProviders();
