+++ src/components/GenericPagesViews/BlogIndex.page.tsx
